% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{addGenes}
\alias{addGenes}
\title{add in gene annotations/information. Also expression information.
Exons = gene - introns
CDS = gene - introns - UTRs
CDS = Exons - UTRs
names(point_in_feature) <- feature_type}
\usage{
addGenes(gtfGR, posGR, ncore = 30)
}
\description{
add in gene annotations/information. Also expression information.
Exons = gene - introns
CDS = gene - introns - UTRs
CDS = Exons - UTRs
names(point_in_feature) <- feature_type
}
