% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{extractCountData}
\alias{extractCountData}
\title{' library(reshape2)
library(foreach)
library(doParallel)
suppressPackageStartupMessages( library( "DEXSeq" ) )}
\usage{
extractCountData(dat, samp.names)
}
\description{
data_list <- data.list
design_vector=c(rep("control",5),rep("treat",5))
names(design_vector) <- names(data_list)
names_vec=row.names(data_list[[1]])
min_samp=3
min_count=1
edits_of_interest=rbind(c("A","G"),c("T","C"))
}
