% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{generateCountFiles}
\alias{generateCountFiles}
\title{colnames(ccounts) <- c("A","T","C","G")
row.names(ccounts) <- common.names
allcounts.treat <- do.call(cbind,counts[which(design_vector=="treat")])
colnames(allcounts.treat) <- paste("treat",1:ncol(allcounts.treat),sep="_")
allcounts.control <- do.call(cbind,counts[which(design_vector=="control")])
colnames(allcounts.control) <- paste("control",1:ncol(allcounts.control),sep="_")
counts.treat <- lapply(allcounts.treat,function(x) data.frame(x,row.names = names_vec))
lapply(counts.treat,nrow)
counts.control <- lapply(counts.control,function(x) data.frame(x[to.use.new,],row.names = to.use.new))
lapply(counts.control,nrow)
print(table(tapply(myexon,paste(mychr,mypos),length)))
head(myname)
head(myv9)
head(newagg)}
\usage{
generateCountFiles(data_list, names_vec = row.names(data_list[[1]]),
  design_vector = c(rep("control", 5), rep("treat", 5)),
  out_dir = "/binf-isilon/alab/projects/ECT2_TC/hyperTRIBE/pipeline/model_tmp_files/")
}
\description{
colnames(ccounts) <- c("A","T","C","G")
row.names(ccounts) <- common.names
allcounts.treat <- do.call(cbind,counts[which(design_vector=="treat")])
colnames(allcounts.treat) <- paste("treat",1:ncol(allcounts.treat),sep="_")
allcounts.control <- do.call(cbind,counts[which(design_vector=="control")])
colnames(allcounts.control) <- paste("control",1:ncol(allcounts.control),sep="_")
counts.treat <- lapply(allcounts.treat,function(x) data.frame(x,row.names = names_vec))
lapply(counts.treat,nrow)
counts.control <- lapply(counts.control,function(x) data.frame(x[to.use.new,],row.names = to.use.new))
lapply(counts.control,nrow)
print(table(tapply(myexon,paste(mychr,mypos),length)))
head(myname)
head(myv9)
head(newagg)
}
