% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plotGeneBody}
\alias{plotGeneBody}
\title{remove
pos_3UTR <- pos_3UTR[!(pos_3UTR %in% c(pos_CDS,pos_5UTR))]}
\usage{
plotGeneBody(myGR, myGR_list = NULL, list_style = F, my.alpha = 0.25,
  my.adjust = 1)
}
\description{
remove
pos_3UTR <- pos_3UTR[!(pos_3UTR %in% c(pos_CDS,pos_5UTR))]
}
