% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{restrict_data}
\alias{restrict_data}
\title{tokeep_list[[i]] <- tokeep_treat & is_max}
\usage{
restrict_data(data_list, design_vector, min_samp = 3, min_count = 0,
  edits_of_interest = rbind(c("A", "G"), c("T", "C")))
}
\description{
tokeep_list[[i]] <- tokeep_treat & is_max
}
